+++ spring-security-rest/src/main/groovy/grails/plugin/springsecurity/rest/token/generation/UUIDTokenGenerator.groovy
@@ -35,7 +35,7 @@ class UUIDTokenGenerator implements TokenGenerator {
     }
 
     @Override
-    AccessToken generateAccessToken(UserDetails principal, Integer expiration) {
+    AccessToken generateAccessToken(UserDetails principal, Long expiration) {
         AccessToken accessToken = generateAccessToken(principal)
         accessToken.expiration = expiration
         return accessToken
